[SDK-parquet] parquet default processor extractor step (#601)
### Description
This is to add an extractor step for parquet default processor.

Changes invovled:
- added a map to filter parquet structs.
- added a new model RawTableItem and a trait `TableItemConvertible` which is implemented for both Parquet and Postgres TableItem. 
- updated a (Postgres) `TableItem::from_write_table_item` to return `table_items` only

Context on why we added a new `RawTableItem` model:
**tl;dr**
This is to reduce the duplicated code as much as possible and reuse the common parsing logic. RawTableItem represents parsed transaction data. This struct will act as the unified output from the parsing function. Parquet TableItem and Postgres TableItem will implement a TableItemConvertible trait to convert RawTableItem into the appropriate format. 

With this approach, parsing logic is centralized, and each processor (Postgres or Parquet) can convert RawTableItem to its respective TableItem type, keeping code DRY and maintainable while supporting format-specific needs.