fix token claims (#627)
## Description 
When a token gets claimed, processor needs to know the previous owner for token_activities_v2. Currently the processor gets it from PendingClaimsResource but in this case the resource wasn't found in the transaction and breaks the processor. This PR tries to get it previous owner from the TokenClaim event instead of the table metadata. 

## Testing
<img width="704" alt="Screenshot 2024-12-06 at 12 44 57 PM" src="https://github.com/user-attachments/assets/ff1dffd7-9c63-424f-8c05-373d671c9077">